*Now validations are made in cities service *Get All Filters made in service *New cities service: find All Active Cities *New provinces service: find All Active Provinces
